Terminación de relaciones tabla review con las funciones mostrar, agregar y eliminar dependiendo de los id de usuario y de película

+++ database/migrations/2021_06_01_022369_create_reviews_table.php
@@ -17,7 +17,7 @@ class CreateReviewsTable extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained();
             $table->foreignId('pelicula_id')->constrained();
-            $table->integer('valoracion');
+            $table->float('valoracion', 2, 2);
             $table->text('resena'); 
             $table->timestamps();
         });
